src/backend/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
imageio==2.31.0
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
moviepy==1.0.3
numpy==1.24.3
orjson==3.8.11
Pillow==9.5.0
proglog==0.1.10
pydantic==1.10.7
PyJWT==1.7.1
pysnowflake==0.1.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
